move code checking if ds is accessible via orga switching to separate service
